rename: 로그인 기능 및 ui 파일 분리(#1)
로그인 기능을 구현하는 메소드는 domain 패키지로, ui를 구성하는 메소드는 screen 패키지로 분리 및 이동했습니다.
