+++ app/src/main/java/com/example/client/MainActivity.kt
@@ -4,10 +4,9 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
+import com.example.client.screen.LoginScreen
 
 class MainActivity : ComponentActivity() {
-    private lateinit var hosts: List<String>
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
