+++ app/src/main/java/com/example/client/screen/LoginScreen.kt
@@ -1,0 +1,28 @@
+package com.example.client.screen
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.platform.LocalContext
+import com.example.client.domain.loginWithKakao
+
+// [todo]: login view 출력
+@Composable
+fun LoginScreen() {
+    val context = LocalContext.current
+
+    Column(
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center,
+    ) {
+        Button(
+            onClick = { loginWithKakao(context) }
+        ) {
+            Text("카카오 로그인")
+        }
+    }
+}
+
